requirements.txt
fastapi==0.115.6
pydantic==2.10.5
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.84.0
oauth2client==4.1.3
vaderSentiment==3.3.2
praw==7.8.1
uvicorn==0.34.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pymongo==4.3.3
email-validator==2.2.0
mangum==0.17.0
httpx==0.28.1
google-cloud-run==0.3.0
gunicorn==21.2.0